Styles refactoring
