+++ cvat-ui/src/base.scss
@@ -1,0 +1,41 @@
+$header-color: #D8D8D8;
+$text-color: #303030;
+$hover-menu-color: rgba(24,144,255,0.05);
+$completed-progress-color: #61C200;
+$inprogress-progress-color: #1890FF;
+$pending-progress-color: #C1C1C1;
+$border-color-1: #c3c3c3;
+$border-color-2: #d9d9d9;
+$border-color-hover: #40a9ff;
+$background-color-1: white;
+$background-color-2: #F1F1F1;
+$transparent-color: rgba(0, 0, 0, 0);
+$player-slider-color: #979797;
+$danger-icon-color: #FF4136;
+$info-icon-color: #0074D9;
+
+$monospaced-fonts-stack: Consolas,Monaco,Lucida Console,Liberation Mono,DejaVu Sans Mono,Bitstream Vera Sans Mono,Courier New, monospace;
+
+hr {
+    border: none;
+    border-top: 1px solid $border-color-1;
+    height: 1px;
+}
+
+.cvat-text-color {
+    color: $text-color;
+}
+
+.cvat-title {
+    font-weight: 400;
+    font-size: 21px;
+    color: $text-color;
+    padding-top: 5px;
+}
+
+#root {
+    width: 100%;
+    height: 100%;
+    min-height: 100%;
+    display: grid;
+}
