+++ cvat-ui/src/components/actions-menu/actions-menu.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import {
