+++ cvat-ui/src/components/annotation-page/annotation-page.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import {
