+++ cvat-ui/src/components/annotation-page/standard-workspace/standard-workspace.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import {
