+++ cvat-ui/src/components/create-model-page/create-model-page.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import {
