+++ cvat-ui/src/components/create-task-page/create-task-page.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import {
