+++ cvat-ui/src/components/file-manager/file-manager.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import {
