+++ cvat-ui/src/components/header/header.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import { RouteComponentProps } from 'react-router';
