+++ cvat-ui/src/components/labels-editor/labels-editor.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import {
