+++ cvat-ui/src/components/model-runner-modal/model-runner-modal.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 
 import {
