+++ cvat-ui/src/components/model-runner-modal/styles.scss
@@ -1,0 +1,13 @@
+@import '../../base.scss';
+
+.cvat-run-model-dialog > div:not(first-child) {
+    margin-top: 10px;
+}
+
+.cvat-run-model-dialog-info-icon {
+    color: $info-icon-color;
+}
+
+.cvat-run-model-dialog-remove-mapping-icon {
+    color: $danger-icon-color;
+}
