+++ cvat-ui/src/components/tasks-page/tasks-page.tsx
@@ -1,3 +1,4 @@
+import './styles.scss';
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { withRouter } from 'react-router-dom';
@@ -15,7 +16,7 @@ import {
 } from '../../reducers/interfaces';
 
 import TopBar from './top-bar';
-import FeedbackComponent from '../feedback';
+import FeedbackComponent from '../feedback/feedback';
 import EmptyListComponent from './empty-list';
 import TaskListContainer from '../../containers/tasks-page/tasks-list';
